requirements.txt
streamlit
langchain
langchain-openai
langchain-chroma
langchain-huggingface
langchain-community
python-dotenv
chromadb
protobuf==3.20.3
sqlite3 >= 3.35.0
sentence-transformers